pll: add hi-res audio PLLs for LFE5U-45/SC3 platform
